Core/Spells: Allow spells with SPELL_DAMAGE_CLASS_NONE to crit
Port From (https://github.com/TrinityCore/TrinityCore/commit/b4064d38c83bc26efa20a248426a9b8cfea6793f)

+++ Source/Game/Entities/Unit/Unit.Fields.cs
@@ -53,7 +53,7 @@ namespace Game.Entities
         public float ModRangedHitChance { get; set; }
         public float ModSpellHitChance { get; set; }
         public bool m_canDualWield;
-        public int BaseSpellCritChance { get; set; }
+        public float BaseSpellCritChance { get; set; }
         public uint RegenTimer { get; set; }
 
         uint _lastExtraAttackSpell;
